fix: streams properly killed if hpack crashes
