+++ test/connection_test.exs
@@ -5,13 +5,8 @@ defmodule Kadabra.ConnectionTest do
     uri = 'http2.golang.org'
     {:ok, pid} = Kadabra.open(uri, :https)
 
-    {_, sup_pid, _, _} =
-      pid
-      |> Supervisor.which_children
-      |> Enum.find(fn({name, _, _, _}) -> name == :connection_sup end)
-
     {_, conn_pid, _, _} =
-      sup_pid
+      pid
       |> Supervisor.which_children
       |> Enum.find(fn({name, _, _, _}) -> name == :connection end)
 
@@ -37,13 +32,8 @@ defmodule Kadabra.ConnectionTest do
       |> Supervisor.which_children
       |> Enum.find(fn({name, _, _, _}) -> name == :stream_sup end)
 
-    {_, conn_sup_pid, _, _} =
-      pid
-      |> Supervisor.which_children
-      |> Enum.find(fn({name, _, _, _}) -> name == :connection_sup end)
-
     {_, conn_pid, _, _} =
-      conn_sup_pid
+      pid
       |> Supervisor.which_children
       |> Enum.find(fn({name, _, _, _}) -> name == :connection end)
